+++ lib/smurff-cpp/SmurffCpp/IO/INIFile.h
@@ -58,9 +58,9 @@ public:
     // returning default_value if not found or not a valid integer (decimal "1234", "-1234", or hex "0x4d2").
     int getInteger(const std::string& section, const std::string& name, int default_value) const;
 
-    // Get a real (floating point double) value from INI file, 
+    // Get a real (floating point float) value from INI file, 
     // returning default_value if not found or not a valid floating point value according to strtod().
-    double getReal(const std::string& section, const std::string& name, double default_value) const;
+    float getReal(const std::string& section, const std::string& name, float default_value) const;
 
     // Get a boolean value from INI file, 
     // returning default_value if not found or if not a valid true/false value.
